requirements.txt
websockets
requests
pyyaml
paho-mqtt
